requirements.txt
# ===========================
# Core Python Requirements
# ===========================
python>=3.8
pandas>=1.3.0
numpy>=1.20.0
scikit-learn>=1.0.0
joblib>=1.0.0
pyyaml>=5.0.0
tqdm>=4.0.0
scipy>=1.7.0
python-box==6.0.2
ensure==1.0.2
types-PyYAML

# ===========================
# NLP Processing
# ===========================
spacy>=3.0.0
fr-core-news-md @ https://github.com/explosion/spacy-models/releases/download/fr_core_news_md-3.7.0/fr_core_news_md-3.7.0-py3-none-any.whl
transformers>=4.0.0
sentencepiece>=0.1.95
tokenizers>=0.10.0

# ===========================
# MLOps Tools
# ===========================
mlflow>=2.2.2
dvc>=3.0.0

# ===========================
# Web & API
# ===========================
Flask
Flask-Cors

# ===========================
# Jupyter & Notebook
# ===========================
notebook
matplotlib

# ===========================
# Deep Learning (optionnel)
# ===========================
torch>=1.10.0  # requis pour les modèles comme CamemBERT
